+++ docs/source/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Latest
+- Implemented automatic versioning of tasks by setting task version to [](#AUTO_VERSION).
+  This feature is currently only supported by Polars [`LazyFrame`](inv:pl#reference/lazyframe/index).
+
 ## 0.6.4 (2023-08-07)
 - Allow invocation of undecorated task functions when calling task object outside of flow definition context.
 - Rename `ignore_fresh_input` to `ignore_cache_function`.